+++ opensquirrel/ir/non_unitary.py
@@ -1,10 +1,10 @@
 from abc import ABC, abstractmethod
-from typing import Any, SupportsInt
+from typing import Any
 
 import numpy as np
 
 from opensquirrel.common import ATOL
-from opensquirrel.ir.expression import Axis, AxisLike, Bit, BitLike, Expression, Int, Qubit, QubitLike
+from opensquirrel.ir.expression import Axis, AxisLike, Bit, BitLike, Expression, Qubit, QubitLike
 from opensquirrel.ir.ir import IRVisitor
 from opensquirrel.ir.statement import Instruction
 
@@ -22,8 +22,8 @@ class NonUnitary(Instruction, ABC):
     def get_qubit_operands(self) -> list[Qubit]:
         return [self.qubit]
 
-    def get_bit_operands(self) -> list[Bit]:
-        return []
+    def accept(self, visitor: IRVisitor) -> Any:
+        return visitor.visit_non_unitary(self)
 
 
 class Measure(NonUnitary):
@@ -43,11 +43,11 @@ class Measure(NonUnitary):
 
     @property
     def arguments(self) -> tuple[Expression, ...]:
-        return self.qubit, self.bit
+        return self.qubit, self.bit, self.axis
 
     def accept(self, visitor: IRVisitor) -> Any:
-        visitor.visit_non_unitary(self)
-        return visitor.visit_measure(self)
+        non_unitary_visit = super().accept(visitor)
+        return non_unitary_visit if non_unitary_visit is not None else visitor.visit_measure(self)
 
     def get_bit_operands(self) -> list[Bit]:
         return [self.bit]
@@ -69,8 +69,8 @@ class Init(NonUnitary):
         return (self.qubit,)
 
     def accept(self, visitor: IRVisitor) -> Any:
-        visitor.visit_non_unitary(self)
-        return visitor.visit_init(self)
+        non_unitary_visit = super().accept(visitor)
+        return non_unitary_visit if non_unitary_visit is not None else visitor.visit_init(self)
 
 
 class Reset(NonUnitary):
@@ -89,46 +89,5 @@ class Reset(NonUnitary):
         return (self.qubit,)
 
     def accept(self, visitor: IRVisitor) -> Any:
-        visitor.visit_non_unitary(self)
-        return visitor.visit_reset(self)
-
-
-class Barrier(NonUnitary):
-    def __init__(self, qubit: QubitLike) -> None:
-        NonUnitary.__init__(self, qubit=qubit, name="barrier")
-        self.qubit = Qubit(qubit)
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}(qubit={self.qubit})"
-
-    def __eq__(self, other: object) -> bool:
-        return isinstance(other, Barrier) and self.qubit == other.qubit
-
-    @property
-    def arguments(self) -> tuple[Expression, ...]:
-        return (self.qubit,)
-
-    def accept(self, visitor: IRVisitor) -> Any:
-        visitor.visit_non_unitary(self)
-        return visitor.visit_barrier(self)
-
-
-class Wait(NonUnitary):
-    def __init__(self, qubit: QubitLike, time: SupportsInt) -> None:
-        NonUnitary.__init__(self, qubit=qubit, name="wait")
-        self.qubit = Qubit(qubit)
-        self.time = Int(time)
-
-    def __repr__(self) -> str:
-        return f"{self.name}(qubit={self.qubit}, time={self.time})"
-
-    def __eq__(self, other: object) -> bool:
-        return isinstance(other, Wait) and self.qubit == other.qubit and self.time == other.time
-
-    @property
-    def arguments(self) -> tuple[Expression, ...]:
-        return self.qubit, self.time
-
-    def accept(self, visitor: IRVisitor) -> Any:
-        visitor.visit_non_unitary(self)
-        return visitor.visit_wait(self)
+        non_unitary_visit = super().accept(visitor)
+        return non_unitary_visit if non_unitary_visit is not None else visitor.visit_reset(self)