+++ opensquirrel/ir/semantics/controlled_gate.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from opensquirrel.ir import Bit, IRVisitor, Qubit, QubitLike
+from opensquirrel.ir import IRVisitor, Qubit, QubitLike
 from opensquirrel.ir.expression import Expression
 from opensquirrel.ir.semantics import BlochSphereRotation
 from opensquirrel.ir.unitary import Gate
@@ -12,8 +12,8 @@ class ControlledGate(Gate):
     ) -> None:
         Gate.__init__(self, name)
         self.control_qubit = Qubit(control_qubit)
-        self.target_qubit = Qubit(target_gate.qubit)
         self.target_gate = target_gate
+        self.target_qubit = Qubit(target_gate.qubit)
 
         if self._check_repeated_qubit_operands([self.control_qubit, self.target_qubit]):
             msg = "control and target qubit cannot be the same"
@@ -23,7 +23,8 @@ class ControlledGate(Gate):
         return f"{self.name}(control_qubit={self.control_qubit}, target_gate={self.target_gate})"
 
     def accept(self, visitor: IRVisitor) -> Any:
-        return visitor.visit_controlled_gate(self)
+        visit_parent = super().accept(visitor)
+        return visit_parent if visit_parent is not None else visitor.visit_controlled_gate(self)
 
     @property
     def arguments(self) -> tuple[Expression, ...]:
@@ -31,9 +32,6 @@ class ControlledGate(Gate):
 
     def get_qubit_operands(self) -> list[Qubit]:
         return [self.control_qubit, self.target_qubit]
-
-    def get_bit_operands(self) -> list[Bit]:
-        return []
 
     def is_identity(self) -> bool:
         return self.target_gate.is_identity()