+++ opensquirrel/ir/semantics/matrix_gate.py
@@ -7,7 +7,7 @@ import numpy as np
 from numpy.typing import ArrayLike, DTypeLike
 
 from opensquirrel.common import ATOL, repr_round
-from opensquirrel.ir import Bit, Qubit, QubitLike
+from opensquirrel.ir import Qubit, QubitLike
 from opensquirrel.ir.unitary import Gate
 
 if TYPE_CHECKING:
@@ -50,13 +50,11 @@ class MatrixGate(Gate):
         return f"{self.name}(qubits={self.operands}, matrix={repr_round(self.matrix)})"
 
     def accept(self, visitor: IRVisitor) -> Any:
-        return visitor.visit_matrix_gate(self)
+        parent_visit = super().accept(visitor)
+        return parent_visit if parent_visit is not None else visitor.visit_matrix_gate(self)
 
     def get_qubit_operands(self) -> list[Qubit]:
         return self.operands
-
-    def get_bit_operands(self) -> list[Bit]:
-        return []
 
     def is_identity(self) -> bool:
         return np.allclose(self.matrix, np.eye(2 ** len(self.operands)), atol=ATOL)