+++ opensquirrel/ir/unitary.py
@@ -1,8 +1,10 @@
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
+from typing import Any
 
 from opensquirrel.common import are_matrices_equivalent_up_to_global_phase
-from opensquirrel.ir.expression import Bit, Expression, Qubit
+from opensquirrel.ir import IRVisitor
+from opensquirrel.ir.expression import Expression, Qubit
 from opensquirrel.ir.statement import Instruction
 
 
@@ -14,11 +16,6 @@ class Unitary(Instruction, ABC):
 class Gate(Unitary, ABC):
     def __init__(self, name: str) -> None:
         Unitary.__init__(self, name)
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, Gate):
-            return False
-        return compare_gates(self, other)
 
     @property
     @abstractmethod
@@ -34,12 +31,16 @@ class Gate(Unitary, ABC):
         pass
 
     @abstractmethod
-    def get_bit_operands(self) -> list[Bit]:
-        pass
-
-    @abstractmethod
     def is_identity(self) -> bool:
         pass
+
+    def accept(self, visitor: IRVisitor) -> Any:
+        return visitor.visit_gate(self)
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, Gate):
+            return False
+        return compare_gates(self, other)
 
 
 def compare_gates(g1: Gate, g2: Gate) -> bool: